+++ src/django_migration_linter/management/utils.py
@@ -46,6 +46,12 @@ def register_linting_configuration_options(parser: CommandParser) -> None:
         help="ignore failures of sqlmigrate command",
     )
 
+    parser.add_argument(
+        "--ignore-initial-migrations",
+        action="store_true",
+        help="ignore initial migrations",
+    )
+
 
 def configure_logging(verbosity: int) -> None:
     logger = logging.getLogger("django_migration_linter")